Improve agent secret regeneration ux
