+++ features/admin.agents.v1/components/edit/agent-credentials.tsx
@@ -17,20 +17,12 @@
  */
 
 import { Grid, Typography } from "@mui/material";
-import { generatePassword, getConfiguration } from "@wso2is/admin.users.v1/utils/generate-password.utils";
-import { useValidationConfigData } from "@wso2is/admin.validation.v1/api";
-import { ValidationFormInterface } from "@wso2is/admin.validation.v1/models/validation-config";
-import { AlertLevels, IdentifiableComponentInterface } from "@wso2is/core/models";
-import { addAlert } from "@wso2is/core/store";
+import { IdentifiableComponentInterface } from "@wso2is/core/models";
 import { Button, CopyInputField, EmphasizedSegment, Message } from "@wso2is/react-components";
-import { AxiosError } from "axios";
-import React, { useMemo } from "react";
+import React from "react";
 import { useTranslation } from "react-i18next";
-import { useDispatch } from "react-redux";
-import { Dispatch } from "redux";
 import { Divider } from "semantic-ui-react";
 import { AgentSecretShowModal } from "./agent-secret-show-modal";
-import { updateAgentPassword } from "../../api/agents";
 
 interface AgentCredentialsProps extends IdentifiableComponentInterface {
     agentId: string;
@@ -40,53 +32,9 @@ export default function AgentCredentials({
     agentId,
     ["data-componentid"]: componentId = "agent-credentials"
 }: AgentCredentialsProps) {
-    const { data: validationData } = useValidationConfigData();
-
-    const dispatch: Dispatch = useDispatch();
     const { t } = useTranslation();
 
-    const passwordValidationConfig: ValidationFormInterface = useMemo((): ValidationFormInterface => {
-        return getConfiguration(validationData);
-    }, [ validationData ]);
-
-    const [ isSecretRegenerationLoading, setIsSecretRegenerationLoading ] = React.useState<boolean>(false);
     const [ isAgentCredentialWizardOpen, setIsAgentCredentialWizardOpen ] = React.useState<boolean>(false);
-    const [ newAgentSecret, setNewAgentSecret ] = React.useState<string>("");
-
-    const regenerateAgentScret = () => {
-        setIsSecretRegenerationLoading(true);
-
-        const newPassword: string = generatePassword(
-            Number(passwordValidationConfig.minLength),
-            Number(passwordValidationConfig.minLowerCaseCharacters) > 0,
-            Number(passwordValidationConfig.minUpperCaseCharacters) > 0,
-            Number(passwordValidationConfig.minNumbers) > 0,
-            Number(passwordValidationConfig.minSpecialCharacters) > 0,
-            Number(passwordValidationConfig.minLowerCaseCharacters),
-            Number(passwordValidationConfig.minUpperCaseCharacters),
-            Number(passwordValidationConfig.minNumbers),
-            Number(passwordValidationConfig.minSpecialCharacters),
-            Number(passwordValidationConfig.minUniqueCharacters)
-        );
-
-        setNewAgentSecret(newPassword);
-
-        updateAgentPassword(
-            agentId,
-            newPassword
-        ).then(() => {
-            setIsAgentCredentialWizardOpen(true);
-        }).catch((_error: AxiosError) => {
-            // Handle error during regeneration of agent secret.
-            dispatch(addAlert({
-                description: t("agents:edit.credentials.regenerate.alerts.error.description"),
-                level: AlertLevels.ERROR,
-                message: t("agents:edit.credentials.regenerate.alerts.error.message")
-            }));
-        }).finally(() => {
-            setIsSecretRegenerationLoading(false);
-        });
-    };
 
     return (
         <EmphasizedSegment padded="very" style={ { border: "none", padding: "21px" } }>
@@ -99,36 +47,29 @@ export default function AgentCredentials({
                     <Typography variant="body1" style={ { marginBottom: "2%" } }>Agent ID</Typography>
                     <CopyInputField value={ agentId } />
                     <Divider />
-                    <Typography variant="body1">Agent secret</Typography>
-                    <Message warning>
-                        <strong>Before regenerating:</strong> Make sure you are ready to update all
-                        applications using the new agent secret. Once regenerated,
-                        <ul>
-                            <li>All scripts and applications that use the current agent secret will
-                                stop working immediately</li>
-                            <li>You will need to update the agent secret in all integrations</li>
-                            <li>The current agent secret cannot be recovered once regenerated</li>
-                        </ul>
+                    <Typography variant="body1">Agent Secret</Typography>
+                    <Message info>
+                        If you’ve lost or forgotten the agent secret, you can regenerate it, but be aware that any
+                        scripts or applications using the current agent secret will need to be updated.
                     </Message>
                     <Button
                         color="red"
-                        loading={ isSecretRegenerationLoading }
-                        disabled={ isSecretRegenerationLoading }
-                        onClick={ regenerateAgentScret }
+                        onClick={ () => {
+                            setIsAgentCredentialWizardOpen(true);
+                        } }
                         data-componentid={ `${componentId}-agent-secret-regenerate-button` }
                     >
-                        { isSecretRegenerationLoading ? "Regenerating" : "Regenerate" }
+                        Regenerate
                     </Button>
                 </Grid>
             </Grid>
             <AgentSecretShowModal
-                title="Regenerated Agent Secret"
                 agentId={ agentId }
-                agentSecret={ newAgentSecret }
                 isOpen={ isAgentCredentialWizardOpen }
                 onClose={ () => {
                     setIsAgentCredentialWizardOpen(false);
                 } }
+                isForSecretRegeneration
             />
         </EmphasizedSegment>
     );
