+++ features/admin.agents.v1/components/edit/agent-secret-show-modal.tsx
@@ -16,10 +16,19 @@
  * under the License.
  */
 
-import { IdentifiableComponentInterface } from "@wso2is/core/models";
-import { CopyInputField } from "@wso2is/react-components";
-import React from "react";
-import { Button, Message, Modal } from "semantic-ui-react";
+import { generatePassword, getConfiguration } from "@wso2is/admin.users.v1/utils/generate-password.utils";
+import { useValidationConfigData } from "@wso2is/admin.validation.v1/api/validation-config";
+import { ValidationFormInterface } from "@wso2is/admin.validation.v1/models/validation-config";
+import { AlertLevels, IdentifiableComponentInterface } from "@wso2is/core/models";
+import { addAlert } from "@wso2is/core/store";
+import { CopyInputField, PrimaryButton } from "@wso2is/react-components";
+import { AxiosError } from "axios";
+import { t } from "i18next";
+import React, { useMemo } from "react";
+import { useDispatch } from "react-redux";
+import { Dispatch } from "redux";
+import { Button, Icon, Message, Modal } from "semantic-ui-react";
+import { updateAgentPassword } from "../../api/agents";
 
 /**
  * Props for the AgentSecretShowModal component.
@@ -28,7 +37,7 @@ interface AgentSecretShowModalProps extends IdentifiableComponentInterface {
     /**
      * The title to be displayed in the modal.
      */
-    title: string;
+    title?: string;
     /**
      * The unique identifier of the agent.
      */
@@ -36,7 +45,7 @@ interface AgentSecretShowModalProps extends IdentifiableComponentInterface {
     /**
      * The secret associated with the agent.
      */
-    agentSecret: string;
+    agentSecret?: string;
     /**
      * Flag indicating whether the modal is open.
      */
@@ -45,6 +54,10 @@ interface AgentSecretShowModalProps extends IdentifiableComponentInterface {
      * Callback to be invoked when the modal is closed.
      */
     onClose: () => void;
+    /**
+     * Is for agent secret regeneration.
+     */
+    isForSecretRegeneration?: boolean;
 }
 
 export function AgentSecretShowModal({
@@ -53,8 +66,55 @@ export function AgentSecretShowModal({
     agentSecret,
     isOpen,
     onClose,
+    isForSecretRegeneration,
     ["data-componentid"]: componentId = "agent-secret-show-modal"
 }: AgentSecretShowModalProps) {
+    const { data: validationData } = useValidationConfigData();
+
+    const dispatch: Dispatch = useDispatch();
+
+    const [ shouldShowSecretRegeneration, setShouldShowSecretRegeneration ] = React.useState<boolean>(false);
+    const [ isSecretRegenerationLoading, setIsSecretRegenerationLoading ] = React.useState<boolean>(false);
+
+    const passwordValidationConfig: ValidationFormInterface = useMemo((): ValidationFormInterface => {
+        return getConfiguration(validationData);
+    }, [ validationData ]);
+
+    const [ newAgentSecret, setNewAgentSecret ] = React.useState<string>("");
+
+    const regenerateAgentScret = () => {
+        setIsSecretRegenerationLoading(true);
+
+        const newPassword: string = generatePassword(
+            Number(passwordValidationConfig.minLength),
+            Number(passwordValidationConfig.minLowerCaseCharacters) > 0,
+            Number(passwordValidationConfig.minUpperCaseCharacters) > 0,
+            Number(passwordValidationConfig.minNumbers) > 0,
+            Number(passwordValidationConfig.minSpecialCharacters) > 0,
+            Number(passwordValidationConfig.minLowerCaseCharacters),
+            Number(passwordValidationConfig.minUpperCaseCharacters),
+            Number(passwordValidationConfig.minNumbers),
+            Number(passwordValidationConfig.minSpecialCharacters),
+            Number(passwordValidationConfig.minUniqueCharacters)
+        );
+
+        setNewAgentSecret(newPassword);
+
+        updateAgentPassword(
+            agentId,
+            newPassword
+        ).then(() => {
+            setShouldShowSecretRegeneration(true);
+        }).catch((_error: AxiosError) => {
+            dispatch(addAlert({
+                description: t("agents:edit.credentials.regenerate.alerts.error.description"),
+                level: AlertLevels.ERROR,
+                message: t("agents:edit.credentials.regenerate.alerts.error.message")
+            }));
+        }).finally(() => {
+            setIsSecretRegenerationLoading(false);
+        });
+    };
 
     return (
         <Modal
@@ -68,47 +128,95 @@ export function AgentSecretShowModal({
             closeOnDimmerClick={ false }
             closeOnEscape
         >
-            <Modal.Header>{ title }</Modal.Header>
+            <Modal.Header>
+                {
+                    title ||
+                    isForSecretRegeneration && !shouldShowSecretRegeneration
+                        ? "Regenerate agent secret"
+                        : "Agent secret regenerated"
+                }</Modal.Header>
             <Modal.Content>
-                <Message warning>
-                    <strong>Important:</strong> Please copy and store the agent credentials securely.{ " " }
+
+                {
+                    isForSecretRegeneration && !shouldShowSecretRegeneration ? (
+                        <>
+                            <strong>Before regenerating:</strong> Make sure you are ready to update all
+                        applications using the new agent secret. Once regenerated,
+                            <ul>
+                                <li>All scripts and applications that use the current agent secret will
+                                stop working immediately.</li>
+                                <li>You will need to update the agent secret in all integrations.</li>
+                                <li>The current agent secret cannot be recovered once regenerated.</li>
+                            </ul>
+                        </>
+                    ) : (<>
+                        <Message warning>
+                            <strong>Important:</strong> Please copy and store the agent credentials securely.{ " " }
                     Make sure to copy the agent secret now as you will not be able to see this again.
-                </Message>
-
-                <label>Agent ID</label>
-                <div style={ { marginTop: "1%" } }>
-                    <CopyInputField
-                        className="agent-id-input"
-                        value={ agentId }
-                        data-componentid="agent-id-readonly-input"
-                    />
-                </div>
-                <div style={ { marginTop: "2%" } }></div>
-                <label>Agent secret</label>
-                <div style={ { marginTop: "1%" } }>
-                    <CopyInputField
-                        className="agent-secret-input"
-                        secret
-                        value={ agentSecret }
-                        hideSecretLabel={ "Hide secret" }
-                        showSecretLabel={ "Show secret" }
-                        data-componentid={ "agent-secret-readonly-input" }
-                    />
-                </div>
+                        </Message>
+                        { !isForSecretRegeneration && (
+                            <>
+                                <label>Agent ID</label>
+                                <div style={ { marginTop: "1%" } }>
+                                    <CopyInputField
+                                        className="agent-id-input"
+                                        value={ agentId }
+                                        data-componentid="agent-id-readonly-input"
+                                    />
+                                </div>
+                            </>
+                        ) }
+                        <div style={ { marginTop: "2%" } }></div>
+                        <label>Agent Secret</label>
+                        <div style={ { marginTop: "1%" } }>
+                            <CopyInputField
+                                className="agent-secret-input"
+                                secret
+                                value={ agentSecret || newAgentSecret }
+                                hideSecretLabel="Hide secret"
+                                showSecretLabel="Show secret"
+                                data-componentid={ "agent-secret-readonly-input" }
+                            />
+                        </div>
+                    </>)
+                }
             </Modal.Content>
-
             <Modal.Actions>
-                <Button
-                    primary={ true }
-                    type="submit"
-                    onClick={ () => {
-                        onClose();
-                    }
-                    }
-                    data-testid={ `${componentId}-confirmation-modal-actions-continue-button` }
-                >
+                { isForSecretRegeneration && !shouldShowSecretRegeneration ? (
+                    <>
+                        <PrimaryButton
+                            className="link-button"
+                            basic
+                            primary
+                            onClick={ () => onClose() }
+                        >
+                        Cancel
+                        </PrimaryButton>
+                        <PrimaryButton
+                            onClick={ () => {
+                                regenerateAgentScret();
+                            } }
+                            loading={ isSecretRegenerationLoading }
+                            disabled={ isSecretRegenerationLoading }
+                        >
+                            Proceed
+                            <Icon name="arrow right" />
+                        </PrimaryButton>
+                    </>
+
+                ) : (
+                    <Button
+                        primary={ true }
+                        type="submit"
+                        onClick={ () => {
+                            onClose();
+                        }
+                        }
+                        data-testid={ `${componentId}-confirmation-modal-actions-continue-button` }
+                    >
                             Done
-                </Button>
+                    </Button>
+                ) }
             </Modal.Actions>
         </Modal>
     );
